Moved shared endpoint to more logic location?
